pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noisechain"
version = "0.1.0"
description = "NoiseChain MVP - Physical Trust Verification Network based on Proof-of-Experience (PoX)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Jung Wook Yang", email = "sadpig70@gmail.com"}
]
keywords = ["noisechain", "pox", "proof-of-experience", "noise-signature", "trust-verification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "psutil>=5.9.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/sadpig70/NoiseChain"
"Repository" = "https://github.com/sadpig70/NoiseChain"
"Issues" = "https://github.com/sadpig70/NoiseChain/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
